requirements.txt
anyio==4.3.0
boto3==1.34.98
botocore==1.34.98
certifi==2024.2.2
click==8.1.7
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jmespath==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.1.1
s3transfer==0.10.1
six==1.16.0
sniffio==1.3.1
urllib3==2.2.1
